requirements.txt
git+https://github.com/creme-ml/creme
django==2.1.7
django-picklefield==2.0
django-rq==1.2.0
jsonfield==2.0.2
psycopg2-binary==2.7.5
redis==2.10.6
requests==2.21.0
rq==0.12.0
rq-scheduler==0.8.3
